move link utils to draft/shared
