+++ src/client/components/draft/paragraph/paragraph.tsx
@@ -6,13 +6,13 @@ import styled from "styled-components"
 import { TextInputUrl } from "../components/text_input_url"
 import { TextNav } from "../components/text_nav"
 import { decorators } from "../shared/decorators"
+import { confirmLink, linkDataFromSelection, removeLink } from "../shared/links"
 import {
   AllowedStylesParagraph,
   StyleMap,
   StyleNamesParagraph,
 } from "../typings"
 import { convertDraftToHtml, convertHtmlToDraft } from "./utils/convert"
-import { confirmLink, linkDataFromSelection, removeLink } from "./utils/links"
 import {
   blockRenderMap,
   handleReturn,
